Add AWS lambda node.js sync and stack scripts
