+++ lambda_functions/execute_stacker_js/serverless.yml
@@ -1,0 +1,66 @@
+# Welcome to Serverless!
+#
+# This file is the main config file for your service.
+#
+# For full config options, check the docs:
+#    docs.serverless.com
+#
+# Happy Coding!
+
+service: execute-stacker
+
+# You can pin your service to only deploy with a specific Serverless version
+# Check out our docs for more details
+# frameworkVersion: "=X.X.X"
+
+# plugins:
+plugins:
+  # install serverless pseudo parameters using the following:
+  #   1) $ npm install serverless-pseudo-parameters --save-dev
+  - serverless-pseudo-parameters
+  # Add simple-ssh as part of plugins so that serverless function shall package this module and deploy to aws
+  - simple-ssh
+
+provider:
+  name: aws
+  runtime: nodejs6.10
+  environment:
+    DEA_MODULE: dea/20180515
+    PROJECT: v10
+    QUEUE: express
+    hostkey: 'orchestrator.raijin.users.default.host'
+    userkey: 'orchestrator.raijin.users.default.user'
+    pkey: 'orchestrator.raijin.users.default.pkey'
+    YEAR: "2017"
+  region: ap-southeast-2
+  stackTags:
+    repo: dea-orchestration
+    author: santosh.mohan@ga.gov.au
+    purpose: nci-automation
+# you can add statements to the Lambda function's IAM Role here
+  iamRoleStatements:
+    - Effect: 'Allow'
+      Action:
+        - 'ssm:GetParameters'
+        - 'ssm:DescribeParameters'
+      Resource:
+        - "arn:aws:ssm:#{AWS::Region}:#{AWS::AccountId}:parameter/orchestrator.*"
+        - "arn:aws:ssm:#{AWS::Region}:#{AWS::AccountId}:parameter/pipeline.*"
+    - Effect: 'Allow'
+      Action: 'kms:Decrypt'
+      Resource:
+        - "arn:aws:kms:#{AWS::Region}:#{AWS::AccountId}:key/eb628cf2-33db-4909-8581-d9c0626ac8ec"
+
+functions:
+  execute_stacker:
+    handler: handler.execute_stacker
+    environment:
+      command: 'execute_stacker --year ${self:provider.environment.YEAR}'
+    events:
+      - schedule:
+          rate: cron(18 00 * * ? *) # Run daily, at 00:15:00 pm UTC time
+          input:
+            dea-module: ${self:provider.environment.DEA_MODULE}
+            project: ${self:provider.environment.PROJECT}
+            queue: ${self:provider.environment.QUEUE}
+            year: ${self:provider.environment.YEAR}
